feat(storage.writestream): increased precision of end stream detection

+++ src/common/abstract/storage.test.ts
@@ -1,14 +1,24 @@
-import {
-    mockFilePath,
-    mockFileContent,
-    mockDir,
-    mockDirContentList,
-    mockSubdir,
-    mockFileStreamPath,
-    mockFileStreamContent,
-    mockSubdirFileName,
-} from '@test/mocks/storage.mock';
+import storageMock from '@test/mocks/storage.mock';
 import { sleep } from '../utils';
+
+const variables: any = {};
+
+const detectCloudName = function (storage) {
+    switch (storage.constructor.name) {
+        case 'S3':
+            return 'aws';
+        case 'Storage':
+            return 'gcp';
+        case 'Fs':
+            return 'local';
+    }
+};
+const getVariables = function (storage) {
+    const cloudName = detectCloudName(storage);
+    if (!cloudName) throw new Error('Cloud name not detected');
+    if (!variables[cloudName]) variables[cloudName] = storageMock(cloudName);
+    return variables[cloudName];
+};
 
 const toBeDefined: any = {};
 toBeDefined.storage = (storage) => {
@@ -45,10 +55,12 @@ createInstance.shouldBeInstanceOf = async (storage, reference) => {
 const sendContent: any = {};
 sendContent.uploadFile = async (storage) => {
     expect.assertions(1);
+    const { mockFilePath, mockFileContent } = getVariables(storage);
     await expect(storage.sendContent(mockFilePath, mockFileContent)).resolves.toBeUndefined();
 };
 sendContent.uploadFileIntoSubDirectory = async (storage) => {
     expect.assertions(1);
+    const { mockDir, mockSubdir, mockSubdirFileName, mockFileContent } = getVariables(storage);
     const _path = [mockDir, mockSubdir, mockSubdirFileName].join('/');
     await expect(storage.sendContent(_path, mockFileContent)).resolves.toBeUndefined();
 };
@@ -56,6 +68,7 @@ sendContent.uploadFileIntoSubDirectory = async (storage) => {
 const readContent: any = {};
 readContent.shouldMatchContent = async (storage) => {
     expect.assertions(1);
+    const { mockFilePath, mockFileContent } = getVariables(storage);
     const value = await storage.readContent(mockFilePath);
     expect(value).toEqual(mockFileContent);
 };
@@ -67,54 +80,56 @@ readContent.shouldThrowErrorForUnexistentFile = async (storage) => {
 const sendStream: any = {};
 sendStream.shouldReturnInstanceOfWriteStream = async (storage, reference) => {
     expect.assertions(1);
+    const { mockFileStreamPath } = getVariables(storage);
     const stream = await storage.sendStream(mockFileStreamPath);
     expect(stream).toBeInstanceOf(reference);
 };
 
-// old method
-// sendStream.shouldSendContent = async (storage) => {
-//     expect.assertions(1);
-//     const writeStream = await storage.sendStream(mockFileStreamPath);
-//     await writeStream.write(mockFileStreamContent);
-
-//     writeStream.on('close', async () => {
-//         const value = await storage.readContent(mockFileStreamPath);
-//         expect(value).toEqual(mockFileStreamContent);
-//     });
-//     await writeStream.end();
-// };
-
-sendStream.shouldSendContent = async (storage, wait = 0) => {
+sendStream.shouldSendShortContent = async (storage) => {
     expect.assertions(1);
-    const stream = await storage.sendStream(mockFileStreamPath);
-    await stream.write(mockFileStreamContent);
+    const { mockFileStreamShortPath, mockFileStreamContent } = getVariables(storage);
+    const stream = await storage.sendStream(mockFileStreamShortPath);
 
-    return new Promise((resolve, reject) => {
-        stream.on('close', async () => {
-            // time for file to be ready on cloud
-            // detected on aws
-            wait && (await sleep(wait));
+    await stream.writeLine(mockFileStreamContent);
+    await stream.end();
 
-            const value = await storage.readContent(mockFileStreamPath);
-            expect(value).toEqual(mockFileStreamContent);
-            resolve(true);
-        });
-        // gcp doesnt trigger finish if "await stream.end()" is used
-        stream.end();
-    });
+    await sendStream.checkFinalContent(storage, mockFileStreamShortPath, mockFileStreamContent);
+};
+
+sendStream.shouldSendLongContent = async (storage) => {
+    expect.assertions(1);
+    const { mockFileStreamLongPath, mockContentLongList } = getVariables(storage);
+    const stream = await storage.sendStream(mockFileStreamLongPath);
+
+    for (const line of mockContentLongList) {
+        await stream.writeLine(line);
+    }
+    await stream.end();
+    const finalContent = mockContentLongList.join('\n');
+
+    await sendStream.checkFinalContent(storage, mockFileStreamLongPath, finalContent);
+};
+
+sendStream.checkFinalContent = async (storage, mockFileStreamPath, finalContent) => {
+    expect.assertions(1);
+    const value = await storage.readContent(mockFileStreamPath);
+    expect(value).toEqual(finalContent);
 };
 
 const readStream: any = {};
 readStream.shouldReturnInstanceOfInterface = async (storage, reference) => {
     expect.assertions(1);
-    const stream = await storage.readStream(mockFileStreamPath);
+    const { mockFileStreamShortPath } = getVariables(storage);
+    const stream = await storage.readStream(mockFileStreamShortPath);
     expect(stream).toBeInstanceOf(reference);
 };
 readStream.shouldMatchContent = async (storage) => {
     expect.assertions(1);
+    const { mockFileStreamShortPath, mockFileStreamContent } = getVariables(storage);
+
     // wait until file is ready for read
-    await sleep(3000);
-    const stream = await storage.readStream(mockFileStreamPath);
+    // await sleep(3000);
+    const stream = await storage.readStream(mockFileStreamShortPath);
     let content = '';
     let firstLine = true;
     for await (const line of stream) {
@@ -134,8 +149,9 @@ readDirectory.shouldHaveContent = async (storage) => {
 };
 readDirectory.shouldMatchContentList = async (storage) => {
     expect.assertions(1);
+    const { mockDir, mockDirContentList } = getVariables(storage);
     const contentList = await storage.readDirectory(mockDir);
-    expect(contentList).toEqual(mockDirContentList);
+    expect(contentList).toEqual(expect.arrayContaining(mockDirContentList));
 };
 readDirectory.shouldHaveNothing = async (storage) => {
     expect.assertions(1);
@@ -151,6 +167,7 @@ getDirectoryContentLength.shouldHaveSomethingIntoRootdir = async (storage) => {
 };
 getDirectoryContentLength.shouldHaveSomethingIntoDir = async (storage) => {
     expect.assertions(1);
+    const { mockDir } = getVariables(storage);
     const value = await storage.getDirectoryContentLength(mockDir);
     expect(value).toBeGreaterThan(0);
 };
@@ -168,6 +185,7 @@ checkDirectoryContentLength.shouldExistRootdir = async (storage) => {
 };
 checkDirectoryContentLength.shouldExistDir = async (storage) => {
     expect.assertions(1);
+    const { mockDir } = getVariables(storage);
     const result = await storage.checkDirectoryContentLength(mockDir);
     expect(result).toBeTruthy();
 };
@@ -177,22 +195,32 @@ checkDirectoryContentLength.shouldNotExist = async (storage) => {
     expect(result).toBeFalsy();
 };
 
+const cleanAfter = true;
 const deleteFile: any = {};
 deleteFile.shouldDo = async (storage) => {
+    if (!cleanAfter) return;
+
     expect.assertions(1);
+    const { mockFilePath } = getVariables(storage);
     await storage.deleteFile(mockFilePath);
     await expect(storage.readContent(mockFilePath)).rejects.toThrow();
 };
 
 const deleteDirectory: any = {};
 deleteDirectory.shouldDeleteRecursively = async (storage) => {
+    if (!cleanAfter) return;
+
     expect.assertions(1);
+    const { mockDir } = getVariables(storage);
     const _path = [mockDir + '/'].join('/');
     await storage.deleteDirectory(_path);
     expect(await storage.checkDirectoryExists(mockDir)).toBeFalsy();
 };
 deleteDirectory.shouldOmitDeletionOfUnexistentDirectory = async (storage) => {
+    if (!cleanAfter) return;
+
     expect.assertions(1);
+    const { mockDir, mockSubdir } = getVariables(storage);
     const _path = [mockDir, mockSubdir + '/'].join('/');
     await storage.deleteDirectory(_path);
     expect(await storage.checkDirectoryExists(mockDir)).toBeFalsy();
